requirements.txt
requests
pandas
finance-datareader
streamlit
google-api-python-client
google-auth-httplib2
google-auth-oauthlib